重构 AutoDownloadProvider 类，移除 MultipleSourceVersionList 类并将其功能整合到 AutoDownloadProvider 中
